coverity#1326953 RV: Bad use of return value
and

coverity#1326954 RV: Bad use of return value
coverity#1326955 RV: Bad use of return value
coverity#1326956 RV: Bad use of return value
coverity#1326957 RV: Bad use of return value
coverity#1326958 RV: Bad use of return value
coverity#1326959 RV: Bad use of return value
coverity#1326960 RV: Bad use of return value
coverity#1326961 RV: Bad use of return value
coverity#1326962 RV: Bad use of return value

+++ qadevOOo/tests/java/mod/_dbaccess/ORowSet.java
@@ -522,8 +522,9 @@ public class ORowSet extends TestCase {
     private final void impl_deleteFile( final String _file )
     {
         java.io.File file = new java.io.File( _file );
-        file.delete();
-        if ( file.exists() )
+
+        boolean bDeleteOk = file.delete();
+        if (!bDeleteOk && file.exists())
             file.deleteOnExit();
     }
